dashboard-estoque-backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.4
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
gunicorn==21.2.0
numpy==2.3.0
openpyxl==3.1.5
oracledb==3.1.1
packaging==25.0
pandas==2.3.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
